suppression lignes inutiles

+++ PEPS/MonteCarloPricer/AsianOption.cpp
@@ -5,7 +5,7 @@ double AsianOption::payoff(const PnlMat* path) const
 {
 	PnlVect* aux = pnl_vect_create(size_);
 	pnl_mat_sum_vect(aux, path, 'r');
-	pnl_vect_div_scalar(aux, (double)nbTimeSteps_ + 1.0);
+	pnl_vect_div_scalar(aux, nbTimeSteps_ + 1.0);
 	double payoff = pnl_vect_scalar_prod(aux, assetWeights_) - strike_;
 	return std::max(payoff, 0.0);
 }
